CV.Arbitrary

Description

This module provides QuickCheck generators for images.

Documentation

smallImage :: Gen (Image GrayScale D32)

Generate a random small image, that might be constant, noisy or smoothly varying Range of values is [0,1]

constImage :: Gen (Image GrayScale D32)

Generate 10x10 constant image

noisyImage :: Gen (Image GrayScale D32)

Generate 10x10 noisy image

smoothImage :: Gen (Image GrayScale D32)

Generate 10x10 smoothly varying image

blockNoise :: Int -> Gen (Image GrayScale D32)

Generate a (10m x 10m) sized noisy image.