CV.Bindings.Types

Documentation

c'IPL_DEPTH_1U :: Num a => a

c'IPL_DEPTH_8U :: Num a => a

c'IPL_DEPTH_16U :: Num a => a

c'IPL_DEPTH_32F :: Num a => a

c'IPL_DEPTH_8S :: Num a => a

c'IPL_DEPTH_16S :: Num a => a

c'IPL_DEPTH_32S :: Num a => a

c'IPL_BORDER_CONSTANT :: Num a => a

c'IPL_BORDER_REPLICATE :: Num a => a

c'IPL_BORDER_REFLECT :: Num a => a

c'IPL_BORDER_WRAP :: Num a => a

data C'IplImage

Constructors

C'IplImage

data C'CvMemStorage

Constructors

C'CvMemStorage

data C'CvSeqBlock

Constructors

C'CvSeqBlock

data C'CvArr

Constructors

C'CvArr

data C'CvHistogram

Constructors

C'CvHistogram

data C'CvSeq

Constructors

C'CvSeq

Fields

c'CvSeq'flags :: CInt

c'CvSeq'header_size :: CInt

c'CvSeq'h_prev :: Ptr C'CvSeq

c'CvSeq'h_next :: Ptr C'CvSeq

c'CvSeq'v_prev :: Ptr C'CvSeq

c'CvSeq'v_next :: Ptr C'CvSeq

c'CvSeq'total :: CInt

c'CvSeq'elem_size :: CInt

c'CvSeq'block_max :: Ptr Char

c'CvSeq'ptr :: Ptr Char

c'CvSeq'delta_elems :: CInt

c'CvSeq'free_blocks :: Ptr C'CvSeqBlock

c'CvSeq'first :: Ptr C'CvSeqBlock

c'extractCVSeq :: Ptr C'CvSeq -> Ptr () -> IO ()

c'cvGetSeqElem :: Ptr C'CvSeq -> CInt -> IO (Ptr CChar)

c'printSeq :: Ptr C'CvSeq -> IO ()

p'extractCVSeq :: FunPtr (Ptr C'CvSeq -> Ptr () -> IO ())

p'cvGetSeqElem :: FunPtr (Ptr C'CvSeq -> CInt -> IO (Ptr CChar))

p'printSeq :: FunPtr (Ptr C'CvSeq -> IO ())

cvSeqToList :: Storable a => Ptr C'CvSeq -> IO [a]

Convert a CvSeq object into list of its contents. Note that since CvSeq can be approximately anything, including a crazy man from the moon, this is pretty unsafe and you must make sure that a is actually the element in the seq, and the seq is something that remotely represents a sequence of elements.

p'CvSeq'flags :: Ptr C'CvSeq -> Ptr CInt

p'CvSeq'header_size :: Ptr C'CvSeq -> Ptr CInt

p'CvSeq'h_prev :: Ptr C'CvSeq -> Ptr (Ptr C'CvSeq)

p'CvSeq'h_next :: Ptr C'CvSeq -> Ptr (Ptr C'CvSeq)

data C'CvRect

Constructors

C'CvRect

Fields

c'CvRect'x :: CInt

c'CvRect'y :: CInt

c'CvRect'width :: CInt

c'CvRect'height :: CInt

p'CvSeq'v_prev :: Ptr C'CvSeq -> Ptr (Ptr C'CvSeq)

p'CvSeq'v_next :: Ptr C'CvSeq -> Ptr (Ptr C'CvSeq)

p'CvSeq'total :: Ptr C'CvSeq -> Ptr CInt

p'CvRect'x :: Ptr C'CvRect -> Ptr CInt

p'CvSeq'elem_size :: Ptr C'CvSeq -> Ptr CInt

p'CvRect'y :: Ptr C'CvRect -> Ptr CInt

p'CvSeq'block_max :: Ptr C'CvSeq -> Ptr (Ptr Char)

p'CvRect'width :: Ptr C'CvRect -> Ptr CInt

p'CvSeq'ptr :: Ptr C'CvSeq -> Ptr (Ptr Char)

p'CvRect'height :: Ptr C'CvRect -> Ptr CInt

p'CvSeq'delta_elems :: Ptr C'CvSeq -> Ptr CInt

p'CvSeq'free_blocks :: Ptr C'CvSeq -> Ptr (Ptr C'CvSeqBlock)

p'CvSeq'first :: Ptr C'CvSeq -> Ptr (Ptr C'CvSeqBlock)

data C'CvScalar

Constructors

C'CvScalar

Fields

c'CvScalar'val_0_ :: CDouble

c'CvScalar'val_1_ :: CDouble

c'CvScalar'val_2_ :: CDouble

c'CvScalar'val_3_ :: CDouble

p'CvScalar'val_0_ :: Ptr C'CvScalar -> Ptr CDouble

p'CvScalar'val_1_ :: Ptr C'CvScalar -> Ptr CDouble

p'CvScalar'val_2_ :: Ptr C'CvScalar -> Ptr CDouble

p'CvScalar'val_3_ :: Ptr C'CvScalar -> Ptr CDouble

data C'CvSize

Constructors

C'CvSize

Fields

c'CvSize'width :: CInt

c'CvSize'height :: CInt

p'CvSize'width :: Ptr C'CvSize -> Ptr CInt

data C'CvSize2D32f

Constructors

C'CvSize2D32f

Fields

c'CvSize2D32f'width :: CFloat

c'CvSize2D32f'height :: CFloat

p'CvSize'height :: Ptr C'CvSize -> Ptr CInt

p'CvSize2D32f'width :: Ptr C'CvSize2D32f -> Ptr CFloat

p'CvSize2D32f'height :: Ptr C'CvSize2D32f -> Ptr CFloat

data C'CvConnectedComp

Constructors

C'CvConnectedComp

Fields

c'CvConnectedComp'area :: CDouble

c'CvConnectedComp'value :: C'CvScalar

c'CvConnectedComp'rect :: C'CvRect

c'CvConnectedComp'contour :: Ptr C'CvSeq

data C'CvPoint

Constructors

C'CvPoint

Fields

c'CvPoint'x :: CInt

c'CvPoint'y :: CInt

p'CvConnectedComp'area :: Ptr C'CvConnectedComp -> Ptr CDouble

p'CvConnectedComp'value :: Ptr C'CvConnectedComp -> Ptr C'CvScalar

p'CvPoint'x :: Ptr C'CvPoint -> Ptr CInt

p'CvConnectedComp'rect :: Ptr C'CvConnectedComp -> Ptr C'CvRect

p'CvPoint'y :: Ptr C'CvPoint -> Ptr CInt

p'CvConnectedComp'contour :: Ptr C'CvConnectedComp -> Ptr (Ptr C'CvSeq)

data C'CvPoint2D32f

Constructors

C'CvPoint2D32f

Fields

c'CvPoint2D32f'x :: CFloat

c'CvPoint2D32f'y :: CFloat

p'CvPoint2D32f'x :: Ptr C'CvPoint2D32f -> Ptr CFloat

p'CvPoint2D32f'y :: Ptr C'CvPoint2D32f -> Ptr CFloat

mkCvPoint2D32F :: (CFloat, CFloat) -> C'CvPoint2D32f

data C'CvBox2D

Constructors

C'CvBox2D

Fields

c'CvBox2D'center :: C'CvPoint2D32f

c'CvBox2D'size :: C'CvSize2D32f

c'CvBox2D'angle :: CFloat

p'CvBox2D'center :: Ptr C'CvBox2D -> Ptr C'CvPoint2D32f

p'CvBox2D'size :: Ptr C'CvBox2D -> Ptr C'CvSize2D32f

p'CvBox2D'angle :: Ptr C'CvBox2D -> Ptr CFloat

data C'CvMoments

Constructors

C'CvMoments

Fields

c'CvMoments'm00 :: CDouble

c'CvMoments'm10 :: CDouble

c'CvMoments'm01 :: CDouble

c'CvMoments'm20 :: CDouble

c'CvMoments'm11 :: CDouble

c'CvMoments'm02 :: CDouble

c'CvMoments'm30 :: CDouble

c'CvMoments'm21 :: CDouble

c'CvMoments'm12 :: CDouble

c'CvMoments'm03 :: CDouble

c'CvMoments'mu20 :: CDouble

c'CvMoments'mu11 :: CDouble

c'CvMoments'mu02 :: CDouble

c'CvMoments'mu30 :: CDouble

c'CvMoments'mu21 :: CDouble

c'CvMoments'mu12 :: CDouble

c'CvMoments'mu03 :: CDouble

c'CvMoments'inv_sqrt_m00 :: CDouble

data C'CvTermCriteria

Constructors

C'CvTermCriteria

Fields

c'CvTermCriteria'type :: CInt

c'CvTermCriteria'max_iter :: CInt

c'CvTermCriteria'epsilon :: Double

data TermCriteria

Constructors

EPS Double

ITER Int

toCvTCrit :: TermCriteria -> C'CvTermCriteria

p'CvTermCriteria'type :: Ptr C'CvTermCriteria -> Ptr CInt

c'CV_TERMCRIT_ITER :: Num a => a

p'CvTermCriteria'max_iter :: Ptr C'CvTermCriteria -> Ptr CInt

c'CV_TERMCRIT_NUMBER :: Num a => a

c'CV_TERMCRIT_EPS :: Num a => a

p'CvTermCriteria'epsilon :: Ptr C'CvTermCriteria -> Ptr Double

c'cvCreateMemStorage :: Int -> IO (Ptr C'CvMemStorage)

c'cvReleaseMemStorage :: Ptr (Ptr C'CvMemStorage) -> IO ()

p'CvMoments'm00 :: Ptr C'CvMoments -> Ptr CDouble

p'cvCreateMemStorage :: FunPtr (Int -> IO (Ptr C'CvMemStorage))

p'cvReleaseMemStorage :: FunPtr (Ptr (Ptr C'CvMemStorage) -> IO ())

withNewMemory :: (Ptr C'CvMemStorage -> IO b) -> IO b

p'CvMoments'm10 :: Ptr C'CvMoments -> Ptr CDouble

p'CvMoments'm01 :: Ptr C'CvMoments -> Ptr CDouble

p'CvMoments'm20 :: Ptr C'CvMoments -> Ptr CDouble

p'CvMoments'm11 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_8UC1 :: Num a => a

c'CV_8UC2 :: Num a => a

p'CvMoments'm02 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_8UC3 :: Num a => a

c'CV_8UC4 :: Num a => a

p'CvMoments'm30 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_8SC1 :: Num a => a

p'CvMoments'm21 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_8SC2 :: Num a => a

c'CV_8SC3 :: Num a => a

p'CvMoments'm12 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_8SC4 :: Num a => a

p'CvMoments'm03 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_16UC1 :: Num a => a

c'CV_16UC2 :: Num a => a

p'CvMoments'mu20 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_16UC3 :: Num a => a

c'CV_16UC4 :: Num a => a

p'CvMoments'mu11 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_16SC1 :: Num a => a

p'CvMoments'mu02 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_16SC2 :: Num a => a

c'CV_16SC3 :: Num a => a

p'CvMoments'mu30 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_16SC4 :: Num a => a

p'CvMoments'mu21 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_32SC1 :: Num a => a

c'CV_32SC2 :: Num a => a

p'CvMoments'mu12 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_32SC3 :: Num a => a

c'CV_32SC4 :: Num a => a

p'CvMoments'mu03 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_32FC1 :: Num a => a

p'CvMoments'inv_sqrt_m00 :: Ptr C'CvMoments -> Ptr CDouble

c'CV_32FC2 :: Num a => a

c'CV_32FC3 :: Num a => a

c'CV_32FC4 :: Num a => a

c'CV_64FC1 :: Num a => a

c'CV_64FC2 :: Num a => a

c'CV_64FC3 :: Num a => a

c'CV_64FC4 :: Num a => a

c'CV_CLOCKWISE :: Num a => a

c'CV_COUNTER_CLOCKWISE :: Num a => a

data C'CvConvexityDefect

Constructors

C'CvConvexityDefect

Fields

c'CvConvexityDefect'start :: Ptr C'CvPoint

c'CvConvexityDefect'end :: Ptr C'CvPoint

c'CvConvexityDefect'depth_point :: Ptr C'CvPoint

c'CvConvexityDefect'depth :: CFloat

p'CvConvexityDefect'start :: Ptr C'CvConvexityDefect -> Ptr (Ptr C'CvPoint)

data C'CvSURFPoint

Constructors

C'CvSURFPoint

Fields

c'CvSURFPoint'pt :: C'CvPoint2D32f

c'CvSURFPoint'laplacian :: CInt

c'CvSURFPoint'size :: CInt

c'CvSURFPoint'dir :: CFloat

c'CvSURFPoint'hessian :: CFloat

p'CvConvexityDefect'end :: Ptr C'CvConvexityDefect -> Ptr (Ptr C'CvPoint)

p'CvConvexityDefect'depth_point :: Ptr C'CvConvexityDefect -> Ptr (Ptr C'CvPoint)

p'CvConvexityDefect'depth :: Ptr C'CvConvexityDefect -> Ptr CFloat

p'CvSURFPoint'pt :: Ptr C'CvSURFPoint -> Ptr C'CvPoint2D32f

p'CvSURFPoint'laplacian :: Ptr C'CvSURFPoint -> Ptr CInt

p'CvSURFPoint'size :: Ptr C'CvSURFPoint -> Ptr CInt

p'CvSURFPoint'dir :: Ptr C'CvSURFPoint -> Ptr CFloat

p'CvSURFPoint'hessian :: Ptr C'CvSURFPoint -> Ptr CFloat