CV.ColourUtils

Description

This module contains functions for simple histogram manipulation. Use this to scale the image for viewing or to perform simple light-level normalization accross multiple images.

Documentation

balance :: (D32, D32) -> Image GrayScale D32 -> Image GrayScale D32

Adjust the image histogram to have fixed mean and standard deviation. This can be used for simple light level normalization.

logarithmicCompression :: Image GrayScale D32 -> Image GrayScale D32

Perform logarithmic compression on the image. This will enhance dark features and suppress bright features. Use this to visualize images with high dynamic range. (FFT results, for example)

stretchHistogram :: Image GrayScale D32 -> Image GrayScale D32

Histogram stretch scales the image to fit the range [0,1]

equalizeHistogram :: Image GrayScale D8 -> Image GrayScale D8

Equalize contrast of the image. This is good for visualizing images with backgrounds and foregrounds that are both bright or both dark.