CV.Conversions

Description

This module provides slow but functional means for exporting images from and to CArrays, which can easily be passed into foreign functions.

Documentation

copyCArrayToImage :: CArray (Int, Int) Double -> Image GrayScale D32

Copy the contents of a CArray into CV.Image type.

copy8UCArrayToImage :: CArray (Int, Int) Word8 -> Image GrayScale D8

Copy the contents of a CArray into CV.Image type.

copyImageToCArray :: Image GrayScale D32 -> CArray (Int, Int) Double

Copy the contents of a CV.Image into a CArray.

copyFCArrayToImage :: CArray (Int, Int) Float -> Image GrayScale D32

Copy CArray of floats to image

copyImageToFCArray :: Image GrayScale D32 -> CArray (Int, Int) Float

Copy D32 grayscale image to CArray

Complex arrays

copyComplexCArrayToImage :: CArray (Int, Int) (Complex Double) -> Image GrayScale D32

Copy the real part of an array to image

copyImageToComplexCArray :: Image GrayScale D32 -> CArray (Int, Int) (Complex Double)

Copy the contents of CV.Image into a pre-existing CArray.

Copy image as a real part of a complex CArray

Copying

copyImageToExistingCArray :: Image t t1 -> CArray i Double -> IO ()

Acquiring images from pointers

unsafe8UC_RGBFromPtr :: (Int, Int) -> Ptr Word8 -> IO (Image RGB D8)

unsafe8UC_BGRFromPtr :: (Int, Int) -> Ptr Word8 -> IO (Image RGB D8)

acquireImageSlowF' :: Int -> Int -> Ptr Float -> IO (Ptr BareImage)

acquireImageSlow' :: Int -> Int -> Ptr Double -> IO (Ptr BareImage)

acquireImageSlow8URGB' :: Int -> Int -> Ptr Word8 -> IO (Ptr BareImage)