CV.Features

Documentation

data SURFParams

Parameters for SURF feature extraction

defaultSURFParams :: SURFParams

Default parameters for getSURF

getSURF

Arguments

:: SURFParams
Method parameters. See defaultSURFParams and mkSURFParams

-> Image GrayScale D8
Input GrayScale image

-> Maybe (Image GrayScale D8)
Optional Binary mask image

-> [(C'CvSURFPoint, [Float])]

Extract Speeded Up Robust Features from an image.

getMSER :: (Point2D a, ELP a ~ Int) => Image GrayScale D8 -> Maybe (Image GrayScale D8) -> MSERParams -> [[a]]

The function encapsulates all the parameters of the MSER extraction algorithm (see http://en.wikipedia.org/wiki/Maximally_stable_extremal_regions

data MSERParams

mkMSERParams

Arguments

:: Int
Delta

-> Int
prune the area which bigger than maxArea

-> Int
prune the area which smaller than minArea

-> Float
prune the area have similar size to its children

-> Float
trace back to cut off mser with diversity < min_diversity

-> Int
for color image, the evolution steps

-> Double
the area threshold to cause re-initialize

-> Double
ignore too small margin

-> Int
the aperture size for edge blur

-> MSERParams

Create parameters for getMSER.

defaultMSERParams :: MSERParams

moments :: Image GrayScale D32 -> Moments

type Moments = C'CvMoments

getSpatialMoment :: (Int, Int) -> Moments -> Double

getCentralMoment :: (Int, Int) -> Moments -> Double

getNormalizedCentralMoment :: (Int, Int) -> Moments -> Double