CV.Filters

Description

This module is a collection of various image filters

Documentation

gaussian :: (Int, Int) -> Image GrayScale D32 -> Image GrayScale D32

gaussianOp :: (Int, Int) -> ImageOperation GrayScale D32

blurOp :: (Int, Int) -> ImageOperation GrayScale D32

blur :: (Int, Int) -> Image GrayScale D32 -> Image GrayScale D32

blurNS :: (Int, Int) -> Image GrayScale D32 -> Image GrayScale D32

class HasMedianFiltering a where

Methods

median :: (Int, Int) -> a -> a

susan :: (Int, Int) -> Double -> Double -> Image GrayScale D32 -> Image GrayScale D32

SUSAN adaptive smoothing filter, see http://users.fmrib.ox.ac.uk/~steve/susan/susan/susan.html

getCentralMoment :: CInt -> (CInt, CInt) -> Image c d -> Image channels depth

getAbsCentralMoment :: CInt -> (CInt, CInt) -> Image c d -> Image channels depth

getMoment :: CInt -> (CInt, CInt) -> Image c d -> Image channels depth

secondMomentBinarize :: CDouble -> Image c d -> Image c d

secondMomentBinarizeOp :: CDouble -> ImageOperation c d

secondMomentAdaptiveBinarize :: CInt -> CInt -> CDouble -> Image c d -> Image c d

secondMomentAdaptiveBinarizeOp :: CInt -> CInt -> CDouble -> ImageOperation c d

selectiveAvg :: (Int, Int) -> Double -> Image GrayScale D32 -> Image GrayScale D32

A selective average filter is an edge preserving noise reduction filter. It is a standard gaussian filter which ignores pixel values that are more than a given threshold away from the filtered pixel value.

convolve2D :: (Point2D anchor, ELP anchor ~ Int) => Matrix D32 -> anchor -> Image GrayScale D32 -> Image GrayScale D32

convolve2DI :: (CInt, CInt) -> Image c1 d1 -> Image c d -> Image channels depth

haar :: IntegralImage -> (Int, Int, Int, Int) -> Image GrayScale D32

Filter the image with box shaped averaging mask.

haarAt :: IntegralImage -> (Int, Int, Int, Int) -> Double

Get an average of a given region.

data IntegralImage

A type for storing integral images. Integral image stores for every pixel the sum of pixels above and left of it. Such images are used for significantly accelerating the calculation of area averages.

integralImage :: Image GrayScale D32 -> IntegralImage

Calculate the integral image from the given image.

verticalAverage :: Image GrayScale D32 -> Image GrayScale D32

Replace pixel values by the average of the row.