CV.Fitting

Documentation

data Ellipse

Constructors

Ellipse

Fields

center :: (Float, Float)

width :: Float

height :: Float

angle :: Float

fitEllipse :: Matrix (Float, Float) -> Ellipse

Given a (1,n) or (n,1) matrix of points, calculate (in the least squares sense) the best ellipse around the points

fitLine2D :: Dist -> Double -> Double -> Double -> Matrix (Float, Float) -> ((Float, Float), (Float, Float))

Fit a line to set of points.

minAreaRect :: Matrix (Float, Float) -> C'CvBox2D

Fit a minimum area rectangle over a set of points

boundingRect :: Matrix (Float, Float) -> C'CvRect

Calculate the minimum axis-aligned bounding rectangle of given points.

boundingCircle :: (ELP a ~ Double, Point2D a) => Matrix (Float, Float) -> (a, Double)

Calculate the minimum enclosing circle of a point set.

convexHull :: Matrix (Float, Float) -> Matrix (Float, Float)

Calculcate the clockwise convex hull of a point set

convexityDefects :: Matrix (Int, Int) -> [(C'CvPoint, C'CvPoint, C'CvPoint, CFloat)]

Calculate convexity defects of a contour.