CV.Histogram

Documentation

newtype HistogramData a

Constructors

HGD [(a, a)]

backProjectHistogram :: [Image GrayScale D8] -> Histogram -> Image GrayScale D8

Given a set of images, such as the color channels of color image, and a histogram with corresponding number of channels, replace the pixels of the image with the likelihoods from the histogram

histogram :: [(Image GrayScale D8, Int)] -> Bool -> Maybe (Image GrayScale D8) -> Histogram

Calculate an opencv histogram object from set of images, each with it's own number of bins.

values :: HistogramData b -> [b]

cmpUnion :: (Num a, Ord a) => [a] -> [a] -> a

cmpIntersect :: (Num a, Ord a) => [a] -> [a] -> a

cmpEuclidian :: Num a => [a] -> [a] -> a

cmpAbs :: Num a => [a] -> [a] -> a

chiSqrHG :: (Fractional a, Ord a) => HistogramData a -> HistogramData a -> a

chiSqr :: (Fractional a, Ord a) => [a] -> [a] -> a

liftBins :: ([b] -> [a]) -> HistogramData b -> [(a, b)]

liftValues :: ([a] -> [b]) -> HistogramData a -> [(a, b)]

sub :: (Eq b, Num b) => HistogramData b -> HistogramData b -> HistogramData b

noBins :: HistogramData t -> Int

getPositivePart :: (Num b, Ord b) => HistogramData b -> HistogramData b

tcumulate :: Num a => [a] -> [a]

weightedHistogram :: (Fractional b, Real a1, Real a) => Image c d -> Image c1 d1 -> a -> a1 -> Int -> [b]

simpleGetHistogram :: Image GrayScale D32 -> Maybe (Image GrayScale D8) -> D32 -> D32 -> Int -> Bool -> [D32]

get_weighted_histogram :: Ptr BareImage -> Ptr BareImage -> CDouble -> CDouble -> CInt -> Ptr CDouble -> IO ()

get_histogram :: Ptr BareImage -> Ptr BareImage -> CFloat -> CFloat -> CInt -> CInt -> Ptr CDouble -> IO ()