CV.ImageMath

Documentation

mkBinaryImageOp :: (Ptr () -> Ptr () -> Ptr () -> IO a) -> Image c1 d1 -> Image c1 d1 -> Image c1 d1

logOp :: ImageOperation GrayScale D32

atan :: Image GrayScale D32 -> Image GrayScale D32

subtractMeanOp :: ImageOperation GrayScale D32

subRSOp :: D32 -> ImageOperation GrayScale D32

mulSOp :: D32 -> ImageOperation GrayScale D32

addSOp :: D32 -> ImageOperation GrayScale D32

mkCmpOp :: CInt -> D32 -> Image GrayScale D32 -> Image GrayScale D8

mkCmp2Op :: CreateImage (Image GrayScale d) => CInt -> Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

moreThan :: D32 -> Image GrayScale D32 -> Image GrayScale D8

lessThan :: D32 -> Image GrayScale D32 -> Image GrayScale D8

lessEq2Than :: CreateImage (Image GrayScale d) => Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

more2Than :: CreateImage (Image GrayScale d) => Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

less2Than :: CreateImage (Image GrayScale d) => Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

average' :: Image GrayScale a -> IO D32

average :: Image GrayScale D32 -> D32

sum :: Image GrayScale D32 -> D32

Sum the pixels in the image. Notice that OpenCV automatically casts the result to double sum :: Image GrayScale D32 -> D32

stdDeviation :: Image GrayScale D32 -> D32

findMinMax' :: BareImage -> BareImage -> IO (D32, D32)

minValue :: Image GrayScale D32 -> D32

Find minimum and maximum value of image i in area specified by the mask.

maxValue :: Image GrayScale D32 -> D32

gaussianImage :: (Int, Int) -> (Double, Double) -> Image GrayScale D32

maskedMerge :: Image GrayScale D8 -> Image GrayScale D32 -> Image GrayScale D32 -> Image GrayScale D32

Produce white image with edgeW amount of edges fading to black

Produce image where pixel is coloured according to distance from the edge

Merge two images according to a mask. Result R is R = A*m+B*(m-1) . TODO: Fix C-code of masked_merge to accept D8 input for the mask

cvAdd :: Ptr () -> Ptr () -> Ptr () -> Ptr () -> IO ()

Given a distance map and a circle, return the biggest circle with radius less than given in the distance map that fully covers the previous one

cvSub :: Ptr () -> Ptr () -> Ptr () -> Ptr () -> IO ()

cvLog :: Ptr () -> Ptr () -> IO ()

sqrtImage :: Ptr BareImage -> Ptr BareImage -> IO ()

cvMin :: Ptr () -> Ptr () -> Ptr () -> IO ()

cvMul :: Ptr () -> Ptr () -> Ptr () -> CDouble -> IO ()

cvDiv :: Ptr () -> Ptr () -> Ptr () -> CDouble -> IO ()

cvMax :: Ptr () -> Ptr () -> Ptr () -> IO ()

cvAbsDiff :: Ptr () -> Ptr () -> Ptr () -> IO ()

calculateAtan :: Ptr BareImage -> Ptr BareImage -> IO ()

calculateAtan2 :: Ptr BareImage -> Ptr BareImage -> Ptr BareImage -> IO ()

wrapAbsDiffS :: Ptr () -> CDouble -> Ptr () -> IO ()

wrapSubRS :: Ptr () -> CDouble -> Ptr () -> IO ()

wrapSubS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvConvertScale :: Ptr () -> Ptr () -> CDouble -> CDouble -> IO ()

wrapAddS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvMinS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvMaxS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvCmpS :: Ptr () -> CDouble -> Ptr () -> CInt -> IO ()

cvCmp :: Ptr () -> Ptr () -> Ptr () -> CInt -> IO ()

wrapAvg :: Ptr () -> IO CDouble

wrapSum :: Ptr () -> IO CDouble

wrapStdDev :: Ptr () -> IO CDouble

wrapStdDevMask :: Ptr () -> Ptr () -> IO CDouble

findMinMax''_ :: Ptr () -> Ptr () -> Ptr CDouble -> Ptr CDouble -> IO ()

wrapMinMaxLoc :: Ptr BareImage -> Ptr CInt -> Ptr CInt -> Ptr CInt -> Ptr CInt -> Ptr CDouble -> Ptr CDouble -> IO ()

render_gaussian :: Ptr BareImage -> CDouble -> CDouble -> IO ()

fadedEdges :: CInt -> CInt -> CInt -> IO (Ptr BareImage)

rectangularDistance :: CInt -> CInt -> IO (Ptr BareImage)

masked_merge :: Ptr BareImage -> Ptr BareImage -> Ptr BareImage -> Ptr BareImage -> IO ()

maximal_covering_circle :: CInt -> CInt -> CDouble -> Ptr BareImage -> Ptr CInt -> Ptr CInt -> Ptr CDouble -> IO ()