CV.ImageMathOp

Description

Mathematical operators for images; see also module ImageMath for the functions these operators are based on.

Operators for two images

(#+), (#*), (#-) :: CreateImage (Image c d) => Image c d -> Image c d -> Image c d

Image addition, subtraction, and multiplication operator; same as add, sub, and mul.

(#<), (#>) :: CreateImage (Image GrayScale d) => Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

Image comparison operators; same as less2Than and more2Than. Example: A #< B produces a binary image that has white pixels in those positions where value of A is less than value of B.

Operators for an image and a scalar

(|*), (-|), (|+) :: D32 -> Image GrayScale D32 -> Image GrayScale D32

Scalar multiplication, addition, and subtraction (scalar on left) operators; same as mulS, addS, and subRS.

(|-) :: Image GrayScale D32 -> D32 -> Image GrayScale D32

Scalar subtraction operator (scalar on right); same as subS.

(|>), (|<) :: D32 -> Image GrayScale D32 -> Image GrayScale D8

Scalar comparison operators; same as moreThan and lessThan. Example: s |> I produces a binary image that has white pixels in those positions where the value of I is larger than s. Notice that this is opposite to the intuitive interpretation.