CV.ImageOp

Documentation

newtype ImageOperation c d

ImageOperation is a name for unary operators that mutate images inplace.

Constructors

ImgOp (Image c d -> IO ())

(#>) :: ImageOperation c d -> ImageOperation c d -> ImageOperation c d

Compose two image operations

nonOp :: ImageOperation c d

An unit operation for compose

(<#) :: Image c d -> ImageOperation c d -> Image c d

Apply image operation to a Copy of an image

fromImageOp :: ImageOperation t t1 -> IOP (Image t t1) (Image t t1)

newtype IOP a b

Constructors

IOP (a -> IO b)

(&#&) :: IOP (Image c d) e -> IOP (Image c d) f -> IOP (Image c d) (Image c d, Image c d)

unsafeOperate :: ImageOperation c d -> Image c d -> Image c d

runIOP :: IOP (Image channels depth) a -> Image channels depth -> IO a

(<##) :: Image c d -> [ImageOperation c d] -> Image c d

Apply list of image operations to a Copy of an image. (Makes a single copy and is faster than folding over (<#)

operate :: ImageOperation c d -> Image c d -> IO (Image c d)

operateOn :: Image c d -> ImageOperation c d -> IO (Image c d)

times :: Int -> ImageOperation c d -> ImageOperation c d

Iterate an operation N times

directOp :: Image t t1 -> ImageOperation t t1 -> IO ()

operateInPlace :: ImageOperation t t1 -> Image t t1 -> IO ()

unsafeOperateOn :: Image c d -> ImageOperation c d -> Image c d

operateWithROI :: (Integral t5, Integral t4, Integral t3, Integral t2) => (t2, t3) -> (t4, t5) -> ImageOperation t t1 -> Image t t1 -> IO (Image t t1)