CV.MultiresolutionSpline

Description

This module provides the elementary image splining (seamless merging) using the burt-adelson multiresolution splines introduced in A multiresolution spline with application to image mosaics, Burt, P.J. and Adelson, E.H., ACM Transactions on Graphics,1983.

Documentation

burtAdelsonMerge :: Int -> Image GrayScale D8 -> Image GrayScale D32 -> Image GrayScale D32 -> Image GrayScale D32

This function merges two images based on given mask, the first image dominates on areas where the mask is 1 and the second where the mask is 0. The merging should be relatively seamless and is controlled by the levels parameter, which adjusts the accuracy. Usually, decent results can be obtained with 4 pyramid levels.

Note that the mask should contain a tiny blurred region between images for optimal result.