CV.Textures

Description

This module provides implementations for basic versions of Local Binary Pattern texture features introduced in T. Ojala, M. Pietikäinen, and D. Harwood (1994), Performance evaluation of texture measures with classification based on Kullback discrimination of distributions, Proceedings of the 12th IAPR International Conference on Pattern Recognition (ICPR 1994).

Documentation

rotationInvariant :: [Double] -> Vector Double

Convert an LBP histogram into rotation invariant form

lbp :: Image GrayScale D32 -> [Double]

The most basic 3x3 lbp operator

lbp5 :: Image GrayScale D32 -> [Double]

The larger radius basic 5x5 lbp operator

weightedLBP :: (Integral a, Integral a1) => a -> a1 -> Image GrayScale D32 -> Image GrayScale D32 -> [Double]

A variant of LBP which is weighted. This can be used to select only parts of the image by using binary masks, or to give higher weight for some areas of the image.