Utils.Stream

Documentation

data Stream m a

Stream of monadic values

Constructors

Terminated

Value (m (a, Stream m a))

sideEffect :: Monad m => (a -> m ()) -> Stream m a -> Stream m a

Attaching side effects

data LRB a b c

Repeating stream

Create a stream by iterating a monadic action

Pure and monadic left fold over a stream

Merge two (time)streams

Constructors

L a

B b

R c

sequenceS :: Monad m => Stream m (m a) -> Stream m a

mapMS :: Monad m => (a -> m b) -> Stream m a -> Stream m b

dropS :: Monad m => Int -> Stream m a -> Stream m a

Drop elements from the stream. Due to stream structure, this operation cannot fail gracefully when dropping more elements than what is found in the stream

takeS :: Monad m => Int -> Stream m a -> Stream m a

pairS :: Monad m => Stream m a -> Stream m (a, a)

terminateOn :: Monad m => (a -> Bool) -> Stream m a -> Stream m a