CV.ImageMath

Documentation

mkBinaryImageOpIO :: (Ptr b -> Ptr b1 -> Ptr b2 -> IO a) -> Image c1 d1 -> Image c d -> IO (Image c1 d1)

mkBinaryImageOp :: (Ptr () -> Ptr () -> Ptr () -> IO a) -> Image c1 d1 -> Image c1 d1 -> Image c1 d1

abcNullPtr :: (t1 -> t2 -> t3 -> Ptr a -> t) -> t1 -> t2 -> t3 -> t

addOp :: Image c1 d1 -> ImageOperation c d

add :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

sub :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

subFrom :: Image c1 d1 -> ImageOperation c d

logOp :: ImageOperation GrayScale D32

log :: Image GrayScale D32 -> Image GrayScale D32

sqrtOp :: ImageOperation c d

sqrt :: Image c d -> Image c d

limitToOp :: Image c1 d1 -> ImageOperation c d

limitTo :: Image c1 d1 -> Image c d -> Image c d

mul :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

div :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

min :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

max :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

absDiff :: Image c1 d1 -> Image c1 d1 -> Image c1 d1

atan :: Image GrayScale D32 -> Image GrayScale D32

atan2 :: CreateImage (Image c2 d2) => Image c d -> Image c1 d1 -> Image c2 d2

subtractMeanAbsOp :: ImageOperation GrayScale D32

invert :: Image GrayScale D32 -> Image GrayScale D32

absOp :: ImageOperation c d

abs :: Image c d -> Image c d

subtractMeanOp :: ImageOperation GrayScale D32

subRSOp :: D32 -> ImageOperation GrayScale D32

subRS :: D32 -> Image GrayScale D32 -> Image GrayScale D32

subSOp :: Real a => a -> ImageOperation c d

subS :: Real a => Image c d -> a -> Image c d

mulSOp :: D32 -> ImageOperation GrayScale D32

mulS :: D32 -> Image GrayScale D32 -> Image GrayScale D32

mkImgScalarOp :: (Fractional b1, Real a1) => (Ptr b -> b1 -> Ptr b -> IO a) -> a1 -> ImageOperation c d

addSOp :: D32 -> ImageOperation GrayScale D32

addS :: D32 -> Image GrayScale D32 -> Image GrayScale D32

minSOp :: Integer -> ImageOperation c d

minS :: Integer -> Image c d -> Image c d

maxSOp :: Integer -> ImageOperation c d

maxS :: Integer -> Image c d -> Image c d

cmpEQ :: Integer

cmpGT :: CInt

cmpGE :: Integer

cmpLT :: CInt

cmpLE :: CInt

cmpNE :: Integer

mkCmpOp :: CInt -> D32 -> Image GrayScale D32 -> Image GrayScale D8

mkCmp2Op :: CreateImage (Image GrayScale d) => CInt -> Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

lessThan, moreThan :: D32 -> Image GrayScale D32 -> Image GrayScale D8

less2Than, more2Than, lessEq2Than :: CreateImage (Image GrayScale d) => Image GrayScale d -> Image GrayScale d -> Image GrayScale D8

averageMask :: Image GrayScale D32 -> Image GrayScale D8 -> D32

average' :: Image GrayScale D32 -> IO D32

average :: Image GrayScale D32 -> D32

sum :: Image GrayScale D32 -> D32

Sum the pixels in the image. Notice that OpenCV automatically casts the result to double sum :: Image GrayScale D32 -> D32

averageImages :: [Image GrayScale D32] -> Image GrayScale D32

stdDeviation' :: Image c d -> IO CDouble

stdDeviation :: Image GrayScale D32 -> D32

stdDeviationMask :: Image c d -> Image c1 d1 -> CDouble

peekFloatConv :: (Storable a, RealFloat a, RealFloat b) => Ptr a -> IO b

findMinMax' :: BareImage -> BareImage -> IO (D32, D32)

findMinMaxLoc :: (Fractional t, Fractional t3, Num t5, Num t4, Num t2, Num t1) => Image c d -> (((t1, t2), t), ((t4, t5), t3))

imageMinMax :: (Fractional t1, Fractional t) => Image c d -> (t, t1)

imageAvgSdv :: (Fractional t7, Fractional t6, Fractional t5, Fractional t4, Fractional t3, Fractional t2, Fractional t1, Fractional t) => Image c d -> ((t, t1, t2, t3), (t4, t5, t6, t7))

findMinMax :: Image t t1 -> (D32, D32)

findMinMaxMask :: BareImage -> BareImage -> (D32, D32)

Find minimum and maximum value of image i in area specified by the mask.

maxValue, minValue :: Image GrayScale D32 -> D32

gaussianImage :: (Int, Int) -> (Double, Double) -> Image GrayScale D32

fadedEdgeImage :: (CInt, CInt) -> CInt -> Image channels depth

Produce white image with edgeW amount of edges fading to black

fadeToCenter :: (CInt, CInt) -> Image channels depth

Produce image where pixel is coloured according to distance from the edge

maskedMerge :: Image GrayScale D8 -> Image GrayScale D32 -> Image GrayScale D32 -> Image GrayScale D32

Merge two images according to a mask. Result R is R = A*m+B*(m-1) . TODO: Fix C-code of masked_merge to accept D8 input for the mask

maximalCoveringCircle :: (Fractional t2, Num t1, Num t) => Image c d -> (CInt, CInt, CDouble) -> (t, t1, t2)

Given a distance map and a circle, return the biggest circle with radius less than given in the distance map that fully covers the previous one

cvAdd :: Ptr () -> Ptr () -> Ptr () -> Ptr () -> IO ()

cvSub :: Ptr () -> Ptr () -> Ptr () -> Ptr () -> IO ()

cvLog :: Ptr () -> Ptr () -> IO ()

sqrtImage :: Ptr BareImage -> Ptr BareImage -> IO ()

cvMin :: Ptr () -> Ptr () -> Ptr () -> IO ()

cvMul :: Ptr () -> Ptr () -> Ptr () -> CDouble -> IO ()

cvDiv :: Ptr () -> Ptr () -> Ptr () -> CDouble -> IO ()

cvMax :: Ptr () -> Ptr () -> Ptr () -> IO ()

cvAbsDiff :: Ptr () -> Ptr () -> Ptr () -> IO ()

calculateAtan :: Ptr BareImage -> Ptr BareImage -> IO ()

calculateAtan2 :: Ptr BareImage -> Ptr BareImage -> Ptr BareImage -> IO ()

wrapAbsDiffS :: Ptr () -> CDouble -> Ptr () -> IO ()

wrapSubRS :: Ptr () -> CDouble -> Ptr () -> IO ()

wrapSubS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvConvertScale :: Ptr () -> Ptr () -> CDouble -> CDouble -> IO ()

wrapAddS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvMinS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvMaxS :: Ptr () -> CDouble -> Ptr () -> IO ()

cvCmpS :: Ptr () -> CDouble -> Ptr () -> CInt -> IO ()

cvCmp :: Ptr () -> Ptr () -> Ptr () -> CInt -> IO ()

wrapAvg :: Ptr () -> Ptr () -> IO CDouble

wrapSum :: Ptr () -> IO CDouble

wrapStdDev :: Ptr () -> IO CDouble

wrapStdDevMask :: Ptr () -> Ptr () -> IO CDouble

findMinMax''_ :: Ptr () -> Ptr () -> Ptr CDouble -> Ptr CDouble -> IO ()

wrapMinMaxLoc :: Ptr BareImage -> Ptr CInt -> Ptr CInt -> Ptr CInt -> Ptr CInt -> Ptr CDouble -> Ptr CDouble -> IO ()

render_gaussian :: Ptr BareImage -> CDouble -> CDouble -> IO ()

fadedEdges :: CInt -> CInt -> CInt -> IO (Ptr BareImage)

rectangularDistance :: CInt -> CInt -> IO (Ptr BareImage)

masked_merge :: Ptr BareImage -> Ptr BareImage -> Ptr BareImage -> Ptr BareImage -> IO ()

maximal_covering_circle :: CInt -> CInt -> CDouble -> Ptr BareImage -> Ptr CInt -> Ptr CInt -> Ptr CDouble -> IO ()