CV.ImageOp

Documentation

newtype ImageOperation c d

ImageOperation is a name for unary operators that mutate images inplace.

Constructors

ImgOp (Image c d -> IO ())

(#>) :: ImageOperation c d -> ImageOperation c d -> ImageOperation c d

Compose two image operations

newtype IOP a b

An unit operation for compose

Apply image operation to a Copy of an image

Constructors

IOP (a -> IO b)

(&#&) :: IOP (Image c d) e -> IOP (Image c d) f -> IOP (Image c d) (Image c d, Image c d)

operate :: ImageOperation c d -> Image c d -> IO (Image c d)

Apply list of image operations to a Copy of an image. (Makes a single copy and is faster than folding over (<#)